+++ holdero/holdero_layout.go
@@ -1,6 +1,7 @@
 package holdero
 
 import (
+	"fmt"
 	"image/color"
 	"time"
 
@@ -19,7 +20,7 @@ import (
 	"fyne.io/fyne/v2/widget"
 )
 
-var H dreams.ContainerStack
+var H dwidget.ContainerStack
 
 // Holdero tables menu tab layout
 func placeContract(change_screen *fyne.Container, d *dreams.AppObject) *fyne.Container {
@@ -141,7 +142,12 @@ func placeContract(change_screen *fyne.Container, d *dreams.AppObject) *fyne.Con
 		}
 	}
 
-	return container.NewStack(bundle.Alpha120, container.NewBorder(contract_cont, nil, nil, nil, tabs))
+	alpha120 := canvas.NewRectangle(color.RGBA{0, 0, 0, 120})
+	if bundle.AppColor == color.White {
+		alpha120 = canvas.NewRectangle(color.NRGBA{R: 0xff, G: 0xff, B: 0xff, A: 0x55})
+	}
+
+	return container.NewStack(alpha120, container.NewBorder(contract_cont, nil, nil, nil, tabs))
 }
 
 // Holdero tab layout
@@ -156,7 +162,7 @@ func placeHoldero(change_screen *widget.Button, d *dreams.AppObject) *fyne.Conta
 		Player6_label(nil, nil, nil),
 		H.TopLabel)
 
-	holdero_label := container.NewHBox(H.LeftLabel, layout.NewSpacer(), H.RightLabel)
+	holdero_label := container.NewHBox(H.Left.Label, layout.NewSpacer(), H.Right.Label)
 
 	H.Front = *placeHolderoCards(d.Window)
 
@@ -184,13 +190,15 @@ func placeHoldero(change_screen *widget.Button, d *dreams.AppObject) *fyne.Conta
 }
 
 // Layout all objects for Holdero dApp
-func LayoutAllItems(d *dreams.AppObject) *fyne.Container {
-	H.LeftLabel = widget.NewLabel("")
-	H.RightLabel = widget.NewLabel("")
+func LayoutAll(d *dreams.AppObject) *fyne.Container {
+	H.Left.Label = widget.NewLabel("")
+	H.Left.Label.SetText("Seats: " + round.display.seats + "      Pot: " + round.display.pot + "      Blinds: " + round.display.blinds + "      Ante: " + round.display.ante + "      Dealer: " + round.display.dealer)
+
+	H.Right.Label = widget.NewLabel("")
+	H.Right.Label.SetText(round.display.readout + "      Player ID: " + round.display.playerId + "      DERO Balance: " + rpc.Wallet.BalanceF("DERO") + "      Height: " + fmt.Sprintf("%d", rpc.Wallet.Height()))
+
 	H.TopLabel = canvas.NewText(round.display.results, color.White)
 	H.TopLabel.Move(fyne.NewPos(387, 204))
-	H.LeftLabel.SetText("Seats: " + round.display.seats + "      Pot: " + round.display.pot + "      Blinds: " + round.display.blinds + "      Ante: " + round.display.ante + "      Dealer: " + round.display.dealer)
-	H.RightLabel.SetText(round.display.readout + "      Player ID: " + round.display.playerId + "      Dero Balance: " + rpc.DisplayBalance("Dero") + "      Height: " + rpc.Wallet.Display.Height)
 
 	var holdero_objs *fyne.Container
 	var contract_objs *fyne.Container