+++ holdero/holdero_rpc.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"time"
 
+	dreams "github.com/dReam-dApps/dReams"
 	"github.com/dReam-dApps/dReams/rpc"
 	"github.com/deroproject/derohe/cryptography/crypto"
 	dero "github.com/deroproject/derohe/rpc"
@@ -261,10 +262,8 @@ func fetchHolderoSC() {
 				getPlayerCardValues(P1C1_jv, P1C2_jv, P2C1_jv, P2C2_jv, P3C1_jv, P3C2_jv, P4C1_jv, P4C2_jv, P5C1_jv, P5C2_jv, P6C1_jv, P6C2_jv)
 			}
 
-			if !rpc.Startup {
-				setHolderoName(OneId_jv, TwoId_jv, ThreeId_jv, FourId_jv, FiveId_jv, SixId_jv)
-				setSignals(Pot_jv, P1Out_jv)
-			}
+			setHolderoName(OneId_jv, TwoId_jv, ThreeId_jv, FourId_jv, FiveId_jv, SixId_jv)
+			setSignals(Pot_jv, P1Out_jv)
 
 			if P1Out_jv != nil {
 				signals.Out1 = true
@@ -276,7 +275,7 @@ func fetchHolderoSC() {
 
 			if FlopBool_jv != nil {
 				round.flop = true
-				rpc.Wallet.KeyLock = false
+				handKeyLock = false
 			} else {
 				round.flop = false
 			}
@@ -352,9 +351,9 @@ func fetchHolderoSC() {
 			if RevealBool_jv != nil && !signals.reveal && !round.localEnd {
 				if rpc.AddOne(Turn_jv) == round.display.playerId {
 					signals.clicked = true
-					signals.height = rpc.Wallet.Height
+					signals.height = rpc.Wallet.Height()
 					signals.reveal = true
-					go RevealKey(rpc.Wallet.ClientKey)
+					go RevealKey(handKey)
 				}
 			}
 
@@ -383,9 +382,9 @@ func fetchHolderoSC() {
 				if round.Last != 0 {
 					now := time.Now().Unix()
 					if now > round.Last+int64(signals.times.kick)+18 {
-						if rpc.Wallet.Height > signals.times.block+3 {
+						if rpc.Wallet.Height() > signals.times.block+3 {
 							TimeOut()
-							signals.times.block = rpc.Wallet.Height
+							signals.times.block = rpc.Wallet.Height()
 						}
 					}
 				}
@@ -404,15 +403,17 @@ func fetchHolderoSC() {
 // Submit playerId, name, avatar and sit at Holdero table
 //   - name and av are for name and avatar in player id string
 func SitDown(name, av string) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	var player playerId
 	player.Id = rpc.Wallet.IdHash
 	player.Name = name
 	player.Avatar = av
 
-	mar, _ := json.Marshal(player)
+	mar, err := json.Marshal(player)
+	if err != nil {
+		rpc.PrintError("[Holdero] Could not make playerID: %s", err)
+		return
+	}
+
 	hx := hex.EncodeToString(mar)
 
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "PlayerEntry"}
@@ -436,7 +437,7 @@ func SitDown(name, av string) (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Sit Down: %s", err)
 		return
 	}
@@ -448,13 +449,10 @@ func SitDown(name, av string) (tx string) {
 
 // Leave Holdero seat on players turn
 func Leave() (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	checkoutId := rpc.StringToInt(round.display.playerId)
 	singleNameClear(checkoutId)
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "PlayerLeave"}
-	arg2 := dero.Argument{Name: "id", DataType: "U", Value: checkoutId}
+	arg2 := dero.Argument{Name: "id", DataType: "U", Value: uint64(checkoutId)}
 	args := dero.Arguments{arg1, arg2}
 	txid := dero.Transfer_Result{}
 
@@ -474,7 +472,7 @@ func Leave() (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Leave: %s", err)
 		return
 	}
@@ -489,16 +487,18 @@ func Leave() (tx string) {
 //   - bb, sb and ante define big blind, small blind and antes. Ante can be 0
 //   - chips defines if tables is using Dero or assets
 //   - name and av are for name and avatar in owners id string
-func SetTable(seats int, bb, sb, ante uint64, chips, name, av string) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
+func SetTable(seats, bb, sb, ante uint64, chips, name, av string) (tx string) {
 	var player playerId
 	player.Id = rpc.Wallet.IdHash
 	player.Name = name
 	player.Avatar = av
 
-	mar, _ := json.Marshal(player)
+	mar, err := json.Marshal(player)
+	if err != nil {
+		rpc.PrintError("[Holdero] Could not make playerID: %s", err)
+		return
+	}
+
 	hx := hex.EncodeToString(mar)
 
 	var args dero.Arguments
@@ -533,7 +533,7 @@ func SetTable(seats int, bb, sb, ante uint64, chips, name, av string) (tx string
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Set Table: %s", err)
 		return
 	}
@@ -545,15 +545,12 @@ func SetTable(seats int, bb, sb, ante uint64, chips, name, av string) (tx string
 
 // Submit blinds/ante to deal Holdero hand
 func DealHand() (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
-	if !rpc.Wallet.KeyLock {
-		rpc.Wallet.ClientKey = generateKey()
+	if !handKeyLock {
+		handKey = generateKey()
 	}
 
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "DealHand"}
-	arg2 := dero.Argument{Name: "pcSeed", DataType: "H", Value: rpc.Wallet.ClientKey}
+	arg2 := dero.Argument{Name: "pcSeed", DataType: "H", Value: crypto.HashHexToHash(handKey)}
 	args := dero.Arguments{arg1, arg2}
 	txid := dero.Transfer_Result{}
 
@@ -608,13 +605,19 @@ func DealHand() (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Deal: %s", err)
 		return
 	}
 
 	round.display.results = ""
 	updateStatsWager(float64(amount) / 100000)
+	if !Odds.Enabled {
+		if err := dreams.StoreAccount(saveAccount()); err != nil {
+			logger.Errorln("[Holdero] storing account", err)
+		}
+	}
+
 	rpc.PrintLog("[Holdero] Deal TX: %s", txid)
 
 	return txid.TXID
@@ -622,9 +625,6 @@ func DealHand() (tx string) {
 
 // Make Holdero bet
 func Bet(amt string) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "Bet"}
 	args := dero.Arguments{arg1}
 	txid := dero.Transfer_Result{}
@@ -662,7 +662,7 @@ func Bet(amt string) (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Bet: %s", err)
 		return
 	}
@@ -680,9 +680,6 @@ func Bet(amt string) (tx string) {
 
 // Holdero check and fold
 func Check() (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "Bet"}
 	args := dero.Arguments{arg1}
 	txid := dero.Transfer_Result{}
@@ -720,7 +717,7 @@ func Check() (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Check/Fold: %s", err)
 		return
 	}
@@ -734,9 +731,6 @@ func Check() (tx string) {
 // Holdero single winner payout
 //   - w defines which player the pot is going to
 func PayOut(w string) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "Winner"}
 	arg2 := dero.Argument{Name: "whoWon", DataType: "S", Value: w}
 	args := dero.Arguments{arg1, arg2}
@@ -758,7 +752,7 @@ func PayOut(w string) (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Payout: %s", err)
 		return
 	}
@@ -771,9 +765,6 @@ func PayOut(w string) (tx string) {
 // Holdero split winners payout
 //   - Pass in ranker from hand and folded bools to determine split
 func PayoutSplit(r ranker, f1, f2, f3, f4, f5, f6 bool) string {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	ways := 0
 	splitWinners := [6]string{"Zero", "Zero", "Zero", "Zero", "Zero", "Zero"}
 
@@ -810,7 +801,7 @@ func PayoutSplit(r ranker, f1, f2, f3, f4, f5, f6 bool) string {
 	sort.Strings(splitWinners[:])
 
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "SplitWinner"}
-	arg2 := dero.Argument{Name: "div", DataType: "U", Value: ways}
+	arg2 := dero.Argument{Name: "div", DataType: "U", Value: uint64(ways)}
 	arg3 := dero.Argument{Name: "split1", DataType: "S", Value: splitWinners[0]}
 	arg4 := dero.Argument{Name: "split2", DataType: "S", Value: splitWinners[1]}
 	arg5 := dero.Argument{Name: "split3", DataType: "S", Value: splitWinners[2]}
@@ -837,7 +828,7 @@ func PayoutSplit(r ranker, f1, f2, f3, f4, f5, f6 bool) string {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Split Winner: %s", err)
 		return ""
 	}
@@ -850,11 +841,9 @@ func PayoutSplit(r ranker, f1, f2, f3, f4, f5, f6 bool) string {
 // Reveal Holdero hand key for showdown
 func RevealKey(key string) {
 	time.Sleep(6 * time.Second)
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
 
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "RevealKey"}
-	arg2 := dero.Argument{Name: "pcSeed", DataType: "H", Value: key}
+	arg2 := dero.Argument{Name: "pcSeed", DataType: "H", Value: crypto.HashHexToHash(key)}
 	args := dero.Arguments{arg1, arg2}
 	txid := dero.Transfer_Result{}
 
@@ -874,7 +863,7 @@ func RevealKey(key string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Reveal: %s", err)
 		return
 	}
@@ -885,9 +874,6 @@ func RevealKey(key string) {
 
 // Owner can shuffle deck for Holdero, clean above 0 can retrieve balance
 func CleanTable(amt uint64) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "CleanTable"}
 	arg2 := dero.Argument{Name: "amount", DataType: "U", Value: amt}
 	args := dero.Arguments{arg1, arg2}
@@ -909,7 +895,7 @@ func CleanTable(amt uint64) (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Clean Table: %s", err)
 		return
 	}
@@ -921,9 +907,6 @@ func CleanTable(amt uint64) (tx string) {
 
 // Owner can timeout a player at Holdero table
 func TimeOut() (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "TimeOut"}
 	args := dero.Arguments{arg1}
 	txid := dero.Transfer_Result{}
@@ -944,7 +927,7 @@ func TimeOut() (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Timeout: %s", err)
 		return
 	}
@@ -956,9 +939,6 @@ func TimeOut() (tx string) {
 
 // Owner can force start a Holdero table with empty seats
 func ForceStat() (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "ForceStart"}
 	args := dero.Arguments{arg1}
 	txid := dero.Transfer_Result{}
@@ -979,7 +959,7 @@ func ForceStat() (tx string) {
 		Fees:      fee,
 	}
 
-	if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+	if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 		rpc.PrintError("[Holdero] Force Start: %s", err)
 		return
 	}
@@ -993,9 +973,6 @@ func ForceStat() (tx string) {
 //   - face and back are the names of assets
 //   - faceUrl and backUrl are the Urls for those assets
 func SharedDeckUrl(face, faceUrl, back, backUrl string) (tx string) {
-	client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-	defer cancel()
-
 	var cards CardSpecs
 	if face != "" && back != "" {
 		cards.Faces.Name = face
@@ -1027,7 +1004,7 @@ func SharedDeckUrl(face, faceUrl, back, backUrl string) (tx string) {
 			Fees:      fee,
 		}
 
-		if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+		if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 			rpc.PrintError("[Holdero] Shared: %s", err)
 			return
 		}
@@ -1043,9 +1020,6 @@ func SharedDeckUrl(face, faceUrl, back, backUrl string) (tx string) {
 // Deposit tournament chip bal with name to leader board SC
 func TourneyDeposit(bal uint64, name string) (tx string) {
 	if bal > 0 {
-		client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-		defer cancel()
-
 		arg1 := dero.Argument{Name: "entrypoint", DataType: "S", Value: "Deposit"}
 		arg2 := dero.Argument{Name: "name", DataType: "S", Value: name}
 		args := dero.Arguments{arg1, arg2}
@@ -1068,7 +1042,7 @@ func TourneyDeposit(bal uint64, name string) (tx string) {
 			Fees:      fee,
 		}
 
-		if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+		if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 			rpc.PrintError("[Holdero] Tournament Deposit: %s", err)
 			return
 		}
@@ -1160,9 +1134,6 @@ func OwnerT3(o bool) (t *dero.Transfer) {
 //   - pub defines public or private SC
 func uploadHolderoContract(pub int) (tx string) {
 	if rpc.IsReady() {
-		client, ctx, cancel := rpc.SetWalletClient(rpc.Wallet.Rpc, rpc.Wallet.UserPass)
-		defer cancel()
-
 		code := GetHolderoCode(pub)
 		if code == "" {
 			rpc.PrintError("[Holdero] Upload: could not get SC code")
@@ -1180,7 +1151,7 @@ func uploadHolderoContract(pub int) (tx string) {
 			Ringsize:  2,
 		}
 
-		if err := client.CallFor(ctx, &txid, "transfer", params); err != nil {
+		if err := rpc.Wallet.CallFor(&txid, "transfer", params); err != nil {
 			rpc.PrintError("[Holdero] Upload: %s", err)
 			return
 		}