+++ holdero/start.go
@@ -2,15 +2,12 @@ package holdero
 
 import (
 	"fmt"
-	"image/color"
 	"os"
 	"os/signal"
 	"runtime"
 	"syscall"
 	"time"
 
-	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/app"
 	"fyne.io/fyne/v2/canvas"
 	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/layout"
@@ -25,9 +22,12 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-const app_tag = "Holdero"
+const (
+	appName = "Holdero"
+	appID   = "dreamdapps.io.holdero"
+)
 
-var version = semver.MustParse("0.3.1-dev.0")
+var version = semver.MustParse("0.3.1-dev.1")
 var gnomon = gnomes.NewGnomes()
 
 // Check holdero package version
@@ -39,35 +39,38 @@ func Version() semver.Version {
 func StartApp() {
 	n := runtime.NumCPU()
 	runtime.GOMAXPROCS(n)
+
+	// Initialize logrus logger to stdout
 	gnomes.InitLogrusLog(logrus.InfoLevel)
-	config := menu.ReadDreamsConfig(app_tag)
-	SetFavoriteTables(config.Tables)
 
-	// Initialize Fyne app and window
-	a := app.NewWithID(fmt.Sprintf("%s Desktop Client", app_tag))
-	a.Settings().SetTheme(bundle.DeroTheme(config.Skin))
-	w := a.NewWindow(app_tag)
-	w.SetIcon(ResourceHolderoIconPng)
-	w.Resize(fyne.NewSize(1400, 800))
-	w.SetMaster()
-	done := make(chan struct{})
+	// Read config.json file
+	config := menu.GetSettings(appName)
 
-	// Initialize dReams AppObject and close func
-	menu.Theme.Img = *canvas.NewImageFromResource(menu.DefaultThemeResource())
-	d := dreams.AppObject{
-		App:        a,
-		Window:     w,
-		Background: container.NewStack(&menu.Theme.Img),
-	}
+	// Set favorites from config
+	// TODO load account elsewhere SetFavoriteTables(config.Tables)
+
+	// Initialize Fyne app and window as dreams.AppObject
+	d := dreams.NewFyneApp(
+		appID,
+		appName,
+		"On-chain Texas Hold'em style poker",
+		bundle.DeroTheme(config.Skin),
+		ResourceHolderoIconPng,
+		menu.DefaultBackgroundResource(),
+		true)
+
+	// Set one channel for Holdero routine
 	d.SetChannels(1)
 	d.SetTab("Holdero")
+
+	// Initialize close func and channel
+	done := make(chan struct{})
 
 	closeFunc := func() {
 		save := dreams.SaveData{
 			Skin:   config.Skin,
 			DBtype: gnomon.DBStorageType(),
-			Tables: GetFavoriteTables(),
-			Theme:  menu.Theme.Name,
+			Theme:  dreams.Theme.Name,
 		}
 
 		if rpc.Daemon.Rpc == "" {
@@ -76,14 +79,14 @@ func StartApp() {
 			save.Daemon = []string{rpc.Daemon.Rpc}
 		}
 
-		menu.WriteDreamsConfig(save)
+		menu.StoreSettings(save)
 		menu.SetClose(true)
-		gnomon.Stop(app_tag)
+		gnomon.Stop(appName)
 		d.StopProcess()
-		w.Close()
+		d.Window.Close()
 	}
 
-	w.SetCloseIntercept(closeFunc)
+	d.Window.SetCloseIntercept(closeFunc)
 
 	// Handle ctrl-c close
 	c := make(chan os.Signal, 1)
@@ -94,7 +97,7 @@ func StartApp() {
 		closeFunc()
 	}()
 
-	// Initialize vars
+	// Initialize Gnomon vars
 	gnomon.SetDBStorageType("boltdb")
 	gnomon.SetFastsync(true, true, 10000)
 
@@ -106,51 +109,48 @@ func StartApp() {
 	form = append(form, widget.NewFormItem("", container.NewVBox(line)))
 	form = append(form, widget.NewFormItem("Avatar", AvatarSelect(menu.Assets.SCIDs)))
 	form = append(form, widget.NewFormItem("Theme", menu.ThemeSelect(&d)))
-	form = append(form, widget.NewFormItem("Card Deck", FaceSelect(menu.Assets.SCIDs)))
+	form = append(form, widget.NewFormItem("Card Deck", FaceSelect(menu.Assets.SCIDs, &d)))
 	form = append(form, widget.NewFormItem("Card Back", BackSelect(menu.Assets.SCIDs)))
-	form = append(form, widget.NewFormItem("Sharing", SharedDecks(&d)))
 	form = append(form, widget.NewFormItem("", layout.NewSpacer()))
 	form = append(form, widget.NewFormItem("", container.NewVBox(line)))
 
-	profile_spacer := canvas.NewRectangle(color.Transparent)
-	profile_spacer.SetMinSize(fyne.NewSize(450, 0))
+	profile := container.NewCenter(container.NewBorder(dwidget.NewSpacer(450, 0), nil, nil, nil, widget.NewForm(form...)))
 
-	profile := container.NewCenter(container.NewBorder(profile_spacer, nil, nil, nil, widget.NewForm(form...)))
+	// Create dwidget connection box, using default OnTapped for RPC/XSWD connections
+	connection := dwidget.NewHorizontalEntries(appName, 1, &d)
 
-	// Create dwidget connection box with controls
-	connect_box := dwidget.NewHorizontalEntries(app_tag, 1)
-	connect_box.Button.OnTapped = func() {
-		rpc.GetAddress(app_tag)
-		rpc.Ping()
-		if rpc.Daemon.IsConnected() && !gnomon.IsInitialized() && !gnomon.IsStarting() {
-			filter := []string{
-				GetHolderoCode(0),
-				GetHolderoCode(2),
-				gnomes.NFA_SEARCH_FILTER,
-				rpc.GetSCCode(rpc.GnomonSCID),
-				rpc.GetSCCode(rpc.RatingSCID),
-				rpc.GetSCCode(rpc.NameSCID)}
+	// Gnomon controlled by daemon connection
+	connection.Connected.OnChanged = func(b bool) {
+		if b {
+			if rpc.Daemon.IsConnected() && !gnomon.IsInitialized() && !gnomon.IsStarting() {
+				filter := []string{
+					GetHolderoCode(0),
+					GetHolderoCode(2),
+					gnomes.NFA_SEARCH_FILTER,
+					rpc.GetSCCode(rpc.GnomonSCID),
+					rpc.GetSCCode(rpc.RatingSCID),
+					rpc.GetSCCode(rpc.NameSCID)}
 
-			go gnomes.StartGnomon(app_tag, gnomon.DBStorageType(), filter, 0, 0, nil)
+				go gnomes.StartGnomon(appName, gnomon.DBStorageType(), filter, 0, 0, nil)
+			}
+		} else {
+			gnomon.Stop(appName)
 		}
 	}
 
-	connect_box.Disconnect.OnChanged = func(b bool) {
-		if !b {
-			gnomon.Stop(app_tag)
-		}
-	}
+	// Set any saved daemon configs
+	connection.AddDaemonOptions(config.Daemon)
 
-	connect_box.AddDaemonOptions(config.Daemon)
-
-	connect_box.Container.Objects[0].(*fyne.Container).Add(menu.StartIndicators())
+	// Adding dReams indicator panel for wallet, daemon and Gnomon
+	connection.AddIndicator(menu.StartIndicators(nil))
 
 	// Layout tabs
 	tabs := container.NewAppTabs(
-		container.NewTabItem(app_tag, LayoutAllItems(&d)),
-		container.NewTabItem("Assets", menu.PlaceAssets(app_tag, profile, nil, ResourceHolderoCirclePng, &d)),
+		container.NewTabItem(appName, LayoutAll(&d)),
+		container.NewTabItem("Assets", menu.PlaceAssets(appName, profile, nil, ResourceHolderoCirclePng, &d)),
+		container.NewTabItem("Market", menu.PlaceMarket(&d)),
 		container.NewTabItem("Swap", PlaceSwap(&d)),
-		container.NewTabItem("Log", rpc.SessionLog(app_tag, version)))
+		container.NewTabItem("Log", rpc.SessionLog(appName, version)))
 
 	tabs.SetTabLocation(container.TabLocationBottom)
 
@@ -163,42 +163,40 @@ func StartApp() {
 			select {
 			case <-ticker.C: // do on interval
 				rpc.Ping()
-				rpc.EchoWallet(app_tag)
-				rpc.GetWalletHeight(app_tag)
-				rpc.GetDreamsBalances(rpc.SCIDs)
+				rpc.Wallet.Sync()
 
-				connect_box.RefreshBalance()
-				if !rpc.Startup {
-					gnomes.EndPoint()
-				}
+				connection.RefreshBalance()
 
-				if rpc.Daemon.IsConnected() && gnomon.IsInitialized() {
-					connect_box.Disconnect.SetChecked(true)
+				if rpc.Daemon.IsConnected() {
+					connection.Connected.SetChecked(true)
 					if gnomon.IsRunning() {
+						gnomes.EndPoint()
 						menu.DisableIndexControls(false)
 						gnomon.IndexContains()
 						menu.Info.RefreshIndexed()
 						if gnomon.HasIndex(2) {
 							gnomon.Checked(true)
 						}
+
+						if gnomon.GetLastHeight() >= gnomon.GetChainHeight()-3 {
+							gnomon.Synced(true)
+						} else {
+							gnomon.Synced(false)
+							gnomon.Checked(false)
+						}
 					}
 
-					menu.Assets.Balances.Refresh()
+					menu.Assets.Balances.List.Refresh()
 					if rpc.Wallet.IsConnected() {
 						menu.Assets.Swap.Show()
 					} else {
 						menu.Assets.Swap.Hide()
 					}
 
-					if gnomon.GetLastHeight() >= gnomon.GetChainHeight()-3 {
-						gnomon.Synced(true)
-					} else {
-						gnomon.Synced(false)
-						gnomon.Checked(false)
-					}
 				} else {
+					gnomon.Synced(false)
 					menu.DisableIndexControls(true)
-					connect_box.Disconnect.SetChecked(false)
+					connection.Connected.SetChecked(false)
 				}
 
 				if !synced && gnomon.IsReady() && rpc.Wallet.Address != "" {
@@ -206,14 +204,10 @@ func StartApp() {
 					synced = true
 				}
 
-				if rpc.Daemon.IsConnected() {
-					rpc.Startup = false
-				}
-
 				d.SignalChannel()
 
 			case <-d.Closing(): // exit
-				logger.Printf("[%s] Closing...", app_tag)
+				logger.Printf("[%s] Closing...", appName)
 				if gnomes.Indicator.Icon != nil {
 					gnomes.Indicator.Icon.Stop()
 				}
@@ -226,11 +220,12 @@ func StartApp() {
 		}
 	}()
 
+	// Start app and place layout
 	go func() {
 		time.Sleep(450 * time.Millisecond)
-		w.SetContent(container.NewStack(d.Background, container.NewStack(bundle.NewAlpha180(), tabs), container.NewVBox(layout.NewSpacer(), connect_box.Container)))
+		d.Window.SetContent(container.NewStack(d.Background, container.NewStack(bundle.NewAlpha180(), tabs), container.NewVBox(layout.NewSpacer(), connection.Container)))
 	}()
-	w.ShowAndRun()
+	d.Window.ShowAndRun()
 	<-done
-	logger.Printf("[%s] Closed\n", app_tag)
+	logger.Printf("[%s] Closed\n", appName)
 }